+++ src/main/java/com/hbm/blocks/bomb/Landmine.java
@@ -164,7 +164,6 @@ public class Landmine extends BlockContainer implements IBomb {
 				ExplosionLarge.spawnShrapnels(world, x + 0.5, y + 0.5, z + 0.5, 5);
 			} else if(this == ModBlocks.mine_fat) {
 				world.spawnEntityInWorld(EntityNukeExplosionMK5.statFac(world, 9, x, y, z)); //31/2
-				EntityNukeTorex.statFac(world, x, y, z, 9); //crude terrorist nuclear weapon 100T
 				//ExplosionNukeSmall.explode(world, x + 0.5, y + 0.5, z + 0.5, ExplosionNukeSmall.PARAMS_MEDIUM);
 
 			}