+++ src/main/java/com/hbm/blocks/bomb/NukeBoy.java
@@ -130,11 +130,6 @@ public class NukeBoy extends BlockContainer implements IBomb {
 			world.spawnEntityInWorld(EntityNukeExplosionMK5.statFac(world, BombConfig.boyRadius, x + 0.5, y + 0.5, z + 0.5));
 			//world.spawnEntityInWorld(EntityNukeCloudSmall.statFac(world, x, y, z, BombConfig.boyRadius));
 
-			EntityNukeTorex torex = new EntityNukeTorex(world);
-			torex.setPositionAndRotation(x + 0.5, y + 1, z + 0.5, 0, 0);
-			torex.getDataWatcher().updateObject(10, 1.5F);
-			world.spawnEntityInWorld(torex);
-			TrackerUtil.setTrackingRange(world, torex, 1000);
 		}
 		return false;
 	}