+++ src/main/java/com/hbm/blocks/bomb/NukeN2.java
@@ -123,7 +123,6 @@ public class NukeN2 extends BlockContainer implements IBomb {
 			world.playSoundEffect(x, y, z, "random.explode", 1.0f, world.rand.nextFloat() * 0.1F + 0.9F);
 
 			world.spawnEntityInWorld(EntityNukeExplosionMK5.statFacNoRad(world, r, x + 0.5, y + 0.5, z + 0.5));
-			EntityNukeTorex.statFac(world, x + 0.5, y + 0.5, z + 0.5, r);
 		}
 
 		return false;