+++ src/main/java/com/hbm/blocks/bomb/NukeShrimp.java
@@ -132,7 +132,6 @@ public class NukeShrimp extends BlockContainer implements IBomb {
 			world.playSoundEffect(x, y, z, "random.explode", 1.0f, world.rand.nextFloat() * 0.1F + 0.9F);
 
 			world.spawnEntityInWorld(EntityNukeExplosionMK5.statFac(world, BombConfig.shrimpRadius, x + 0.5, y + 0.5, z + 0.5));
-			EntityNukeTorex.statFac(world, x + 0.5, y + 0.5, z + 0.5, BombConfig.shrimpRadius);
 		}
 
 		return false;