+++ src/main/java/com/hbm/main/ModEventHandlerClient.java
@@ -161,10 +161,13 @@ import static com.hbm.explosion.ExplosionNukeSmall.PARAMS_VISUALNOSHRAP;
 
 public class ModEventHandlerClient {
 
-	public static final int flashDuration = 5_000;
+	public static int flashDuration = 5_000;
 	public static long flashTimestamp;
 	public static final int shakeDuration = 1_500;
 	public static long shakeTimestamp;
+	public static long seismicTimestamp;
+	public static int seismicDuration;
+	public static float seismicIntensity;
 
 	@SubscribeEvent
 	public void onOverlayRender(RenderGameOverlayEvent.Pre event) {